refactor: open login modal
