+++ src/components/Actions/Like.tsx
@@ -11,9 +11,9 @@ import { classNames } from '@/helpers/classNames.js';
 import { enqueueErrorMessage, enqueueSuccessMessage } from '@/helpers/enqueueMessage.js';
 import { nFormatter } from '@/helpers/formatCommentCounts.js';
 import { getSnackbarMessageFromError } from '@/helpers/getSnackbarMessageFromError.js';
+import { openLoginModal } from '@/helpers/openLoginModal.js';
 import { resolveSocialMediaProvider } from '@/helpers/resolveSocialMediaProvider.js';
 import { useIsLogin } from '@/hooks/useIsLogin.js';
-import { LoginModalRef } from '@/modals/controls.js';
 
 interface LikeProps {
     postId: string;
@@ -42,7 +42,7 @@ export const Like = memo<LikeProps>(function Like({
         if (!postId) return null;
 
         if (!isLogin) {
-            LoginModalRef.open({ source });
+            openLoginModal({ source });
             return;
         }
