+++ src/components/Compose/PostByItem.tsx
@@ -14,11 +14,12 @@ import { classNames } from '@/helpers/classNames.js';
 import { enqueueErrorMessage, enqueueSuccessMessage } from '@/helpers/enqueueMessage.js';
 import { getSnackbarMessageFromError } from '@/helpers/getSnackbarMessageFromError.js';
 import { isSameProfile } from '@/helpers/isSameProfile.js';
+import { openLoginModal } from '@/helpers/openLoginModal.js';
 import { resolveSourceName } from '@/helpers/resolveSourceName.js';
 import { useAccounts } from '@/hooks/useAccounts.js';
 import { useCompositePost } from '@/hooks/useCompositePost.js';
 import { useCurrentProfile } from '@/hooks/useCurrentProfile.js';
-import { ComposeModalRef, LoginModalRef } from '@/modals/controls.js';
+import { ComposeModalRef } from '@/modals/controls.js';
 import type { Account } from '@/providers/types/Account.js';
 import { useComposeStateStore } from '@/store/useComposeStore.js';
 
@@ -65,7 +66,7 @@ export function PostByItem({ source, disabled = false }: PostByItemProps) {
 
                             ComposeModalRef.close();
                             await delay(300);
-                            LoginModalRef.open({
+                            openLoginModal({
                                 source,
                             });
                         }}
