+++ src/components/Login/LoginStatusBar.tsx
@@ -8,8 +8,8 @@ import { ProfileAvatarInteractive } from '@/components/Login/ProfileAvatarIntera
 import { type SocialSource } from '@/constants/enum.js';
 import { SORTED_SOCIAL_SOURCES } from '@/constants/index.js';
 import { classNames } from '@/helpers/classNames.js';
+import { openLoginModal } from '@/helpers/openLoginModal.js';
 import { useCurrentProfileAll } from '@/hooks/useCurrentProfile.js';
-import { LoginModalRef } from '@/modals/controls.js';
 import { useNavigatorState } from '@/store/useNavigatorStore.js';
 
 interface LoginStatusBarProps {
@@ -24,7 +24,7 @@ export function LoginStatusBar({ collapsed = false }: LoginStatusBarProps) {
     const [{ loading }, onLogin] = useAsyncFn(async (source: SocialSource) => {
         updateSidebarOpen(false);
         await delay(300);
-        LoginModalRef.open({ source });
+        openLoginModal({ source });
     }, []);
 
     return (
