+++ src/components/Poll/VoteButton.tsx
@@ -5,9 +5,9 @@ import LoadingIcon from '@/assets/loading.svg';
 import { ClickableButton } from '@/components/ClickableButton.js';
 import { enqueueErrorMessage, enqueueSuccessMessage } from '@/helpers/enqueueMessage.js';
 import { getSnackbarMessageFromError } from '@/helpers/getSnackbarMessageFromError.js';
+import { openLoginModal } from '@/helpers/openLoginModal.js';
 import { resolvePollProvider } from '@/helpers/resolvePollProvider.js';
 import { useIsLogin } from '@/hooks/useIsLogin.js';
-import { LoginModalRef } from '@/modals/controls.js';
 import type { PollOption } from '@/providers/types/Poll.js';
 import type { Post } from '@/providers/types/SocialMedia.js';
 
@@ -24,7 +24,7 @@ export function VoteButton({ option, post, frameUrl }: VoteButtonProps) {
     const [{ loading }, handleVote] = useAsyncFn(async () => {
         try {
             if (!isLogin) {
-                LoginModalRef.open({ source: post.source });
+                openLoginModal({ source: post.source });
                 return;
             }
             if (!pollId) throw new Error('Poll ID not found');
