+++ src/helpers/checkers.ts
@@ -1,10 +1,10 @@
 import { FarcasterSignType, Source } from '@/constants/enum.js';
 import { FarcasterInvalidSignerKey } from '@/constants/error.js';
-import { LoginModalRef } from '@/modals/controls.js';
+import { openLoginModal } from '@/helpers/openLoginModal.js';
 
 export function checkFarcasterInvalidSignerKey(error: unknown) {
     if (error instanceof FarcasterInvalidSignerKey) {
-        LoginModalRef.open({
+        openLoginModal({
             source: Source.Farcaster,
             options: {
                 expectedSignType: FarcasterSignType.GrantPermission,
