+++ src/hooks/useToggleBookmark.ts
@@ -4,9 +4,9 @@ import { useMutation } from '@tanstack/react-query';
 import { BookmarkType, type SocialSource } from '@/constants/enum.js';
 import { enqueueErrorMessage, enqueueSuccessMessage } from '@/helpers/enqueueMessage.js';
 import { getSnackbarMessageFromError } from '@/helpers/getSnackbarMessageFromError.js';
+import { openLoginModal } from '@/helpers/openLoginModal.js';
 import { resolveSocialMediaProvider } from '@/helpers/resolveSocialMediaProvider.js';
 import { useIsLogin } from '@/hooks/useIsLogin.js';
-import { LoginModalRef } from '@/modals/controls.js';
 import type { Post } from '@/providers/types/SocialMedia.js';
 
 export function useToggleBookmark(source: SocialSource) {
@@ -15,7 +15,7 @@ export function useToggleBookmark(source: SocialSource) {
         mutationFn: async (post: Post) => {
             const { hasBookmarked, postId } = post;
             if (!isLogin) {
-                LoginModalRef.open({ source: post.source });
+                openLoginModal({ source: post.source });
                 return;
             }
             try {
