+++ src/locales/ja/messages.po
@@ -2198,4 +2198,3 @@ msgstr ""
 
 msgid "Blink Validator"
 msgstr ""
-
