+++ src/locales/zh-Hans/messages.po
@@ -2198,4 +2198,3 @@ msgstr "× {0}"
 
 msgid "Blink Validator"
 msgstr "Blink 验证器"
-
